# openapi.yaml
openapi: 3.0.3
info:
  title: Social Media API
  version: 1.0.0
  description: |
    REST API for a simple social media app with AI trip planner.
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
  /plan-trip:
    post:
      summary: AI Trip Planner (powered by Dockerized Ollama/phi3.mini)
      operationId: planTrip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripPlanRequest'
      responses:
        '200':
          description: Trip plan generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TripPlanResponse'
  /posts:
    post:
      summary: Create a new post
      operationId: createPost
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreate'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
    get:
      summary: List or search posts
      operationId: getOrSearchPosts
      parameters:
        - name: search
          in: query
          description: Returns the latest 20 posts by default, ordered by creation time (descending). If a search term is provided, filters posts by matching username or body.
          required: false
          schema:
            type: string
        - name: limit
          in: query
          description: Max number of posts to return (default 20)
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          description: Number of posts to skip (for pagination)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
      responses:
        '200':
          description: List of posts (filtered if search term is given)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'
  /images/{image_id}:
    get:
      summary: Get full-size image
      operationId: getImage
      parameters:
        - name: image_id
          in: path
          required: true
          description: UUID of the image
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Full-size image
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
        '404':
          description: Image not found
  /images/{image_id}/thumbnail:
    get:
      summary: Get thumbnail version of image
      operationId: getImageThumbnail
      description: Returns a resized thumbnail (max 400px width). If thumbnail not yet generated, returns full image with Cache-Control no-store.
      parameters:
        - name: image_id
          in: path
          required: true
          description: UUID of the image
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Thumbnail image (or full image if thumbnail not ready)
          headers:
            X-Is-Thumbnail:
              schema:
                type: boolean
              description: Whether this is a real thumbnail (true) or fallback full image (false)
            Cache-Control:
              schema:
                type: string
              description: 'no-store if fallback, immutable if real thumbnail'
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
        '404':
          description: Image not found


components:
  schemas:
    TripPlanRequest:
      type: object
      required:
        - country
        - concept
        - budget
        - days
      properties:
        country:
          type: string
          example: "Italy"
        concept:
          type: string
          example: "food and culture"
        budget:
          type: string
          example: "2000"
        days:
          type: string
          example: "7"
        people:
          type: string
          example: "2"
    TripPlanResponse:
      type: object
      properties:
        plan:
          type: string
          example: |
            Day 1: Arrive in Rome, visit the Colosseum...
    PostCreate:
      type: object
      required:
        - username
        - body
      properties:
        username:
          type: string
          example: "Joe King"
        body:
          type: string
          example: "Totally #nofilter"
        image_path:
          type: string
          nullable: true
          example: "./Testdata/p3.png"
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        body:
          type: string
        image_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
 